fix(dm): derive group bubble wireKind (14 chat / 15 file); split MessageBubble footer
Copilot blocking review (#953): GroupConversationScreen hard-coded
wireKind={14} for every group bubble, but group sends can be kind-15
encrypted file messages (useGroupMessaging → createGroupFileRumor). The
info sheet therefore misreported the kind and, worse, left Re-publish
enabled for kind-15 files (useMessageInfoSheet gates canResend on
wireKind === 14).

- Add deriveGroupWireKind(text): returns 15 when the stored text is an
  encrypted #lpe=1 file payload (voice/image, matching createGroupFileRumor),
  else 14 — mirroring the send-side split in useGroupComposerActions.
- Precompute per-message wireKind in classifiedMessages and pass it through
  instead of a literal 14, so file bubbles report kind 15 and correctly
  disable Re-publish.
- Unit-test deriveGroupWireKind (encrypted voice/image → 15; chat text,
  plain media URL, and malformed lpe fragment → 14).

Also fixes the File size gate: MessageBubble.tsx had crept to 1005 lines
(over the 1000 cap). Extract the cohesive DeliveryTick + BubbleFooter pair
into src/components/MessageBubbleFooter.tsx (DeliveryTick was only used by
BubbleFooter), bringing MessageBubble.tsx to 857 lines. No behaviour change.
