+++ src/screens/GroupConversationScreen.tsx
@@ -47,6 +47,7 @@ import { fetchProfile, DEFAULT_RELAYS } from '../services/nostrService';
 import { loadGroupMessages, type GroupMessage } from '../services/groupMessagesStorageService';
 import {
   classifyMessageContent,
+  deriveGroupWireKind,
   extractSharedContact,
   type BubbleContent,
 } from '../utils/messageContent';
@@ -62,7 +63,7 @@ type GroupConversationNavigation = NativeStackNavigationProp<
 
 // Pre-classified variant of GroupMessage — created in a useMemo so
 // classifyMessageContent is NOT called inside the hot renderMessage path.
-type ClassifiedMessage = GroupMessage & { content: BubbleContent };
+type ClassifiedMessage = GroupMessage & { content: BubbleContent; wireKind: 14 | 15 };
 
 interface MemberRow {
   pubkey: string;
@@ -373,7 +374,15 @@ const GroupConversationScreen: React.FC = () => {
     // Sanitise before classify so a tofu placeholder (#764) never reaches
     // the bubble's text branch.
     () =>
-      messages.map((m) => ({ ...m, content: classifyMessageContent(sanitizeDisplayText(m.text)) })),
+      messages.map((m) => ({
+        ...m,
+        content: classifyMessageContent(sanitizeDisplayText(m.text)),
+        // Derive the real NIP-17 kind (14 chat / 15 encrypted file) from the
+        // stored text rather than hard-coding 14, so the info sheet reports
+        // kind-15 for voice/image file bubbles. Precomputed here (not in the
+        // hot renderMessage path) alongside the content classification.
+        wireKind: deriveGroupWireKind(m.text),
+      })),
     [messages],
   );
 
@@ -414,10 +423,12 @@ const GroupConversationScreen: React.FC = () => {
           onOpenGifFullscreen={setFullscreenGifUrl}
           onToggleSecretMode={handleToggleSecretMode}
           isInvoicePaid={isInvoicePaid}
-          // Group DMs are NIP-17 gift-wrapped kind-14 — the info sheet reads the
-          // protocol/kind off this. Group sends aren't per-relay tracked, so no
-          // deliveryStatus (the sheet shows "Not tracked").
-          wireKind={14}
+          // Group DMs are NIP-17 gift-wrapped: kind-14 chat or kind-15 encrypted
+          // file (voice/image). The info sheet reads the protocol/kind off this,
+          // so pass the per-message wireKind derived above rather than a hard
+          // 14. Group sends aren't per-relay tracked, so no deliveryStatus (the
+          // sheet shows "Not tracked").
+          wireKind={item.wireKind}
           onShowInfo={handleShowInfo}
           testIdPrefix="group-conversation"
         />
